Improved EtherCAT rescan performance.
The SII data and PDOs will now be re-used after a re-scan when the
EtherCAT master is in its operation phase. The SII data and PDOs will be
detached from the slaves prior to a scanning and re-attached during the
scanning without the need to fetch the SII data and PDOs once again.
The SII data and PDOs will however only be re-used if the slave can
match either the alias or the combination of vendor id, product code and
serial number for some already stored SII data.
* * *
Fix to ensure that master bails out if no SII data is available. Master
could previously get stucked in the same state in fsm_slave_config
and spam syslog.