Fix accidentially remove of Makefile
The call `make mrproper` removes the Makefile template for building a valid
Synapticon release of the IGH master. Since makefile are usually generated by
autohell and I coudln't find a easy way for a exception I just added a
Makefile.template which is copied to Makefile before the release archive is
generated.

Nevertheless, the Makefile should be replaced by a shell script afeter all.