Increment interface versions.
This is a placeholder assuming that you will be applying patches that change
the ABI for both ecrt and ioctl interfaces.  It should not be applied upstream
as-is, and if you are only applying subsets of patches at a time (and you want
to know when you have mismatched library and app code and need to recompile)
then you may want to increment the versions as you apply different sets.