Reduce waiting period for SDOs
It appears that we are still uploading SDOs when the master tries to startup.
That leaves us without all the SDOs uploaded. I am removing the delay between
the time that the slave enters PREOP and the time the SDOs are uploaded to see
if it reduces the frequency of "max activation" errors.