Added ability to explicity add and remove eoe interfaces so that they can be pre-added before the slaves are scanned and also added if the slave is not yet connected.  Manually configured eoe interfaces now also remain configured if the slave is disconnected.  Added tool commands eoe_addif and eoe_delif.  Updated tool to allow exact match for commands.
EOE interfaces can now be configured at module load time
("eoe_interfaces" parameter) and also whether to allow auto
or manual only creation of eoe interfaces
("eoe_autocreate" parameter)

EOE slaves no longer being kept in op mode when deactivating
the master.  The eoe port still keeps running anyway and your
app may be waiting for all slaves to be in PREOP before exiting

Fixed "BUG: scheduling while atomic" error with EOE transmit
queue.  ec_eoedev_tx() runs in an interrupt context and is not
allowed to sleep, so it is not allowed to use a semaphore to
protect the queue.  The queue has been replaced with a ring
buffer.