Fix the memory leak when fetching a slave config
The ecrt_master_slave_config function keeps creating a new  slave config struct on each call, which should not happen. The correct behavior should be that the ecrt_master_slave_config function would create a new slave config only if it doesn't already exist for the used slave, or if it does exist, just return a pointer to it.