Properly free the process data memory when deactivating the master
Since the process data memory is not being properly freed during master deactivation, reactivating the master with a different sized process data, without completely deinitializing it first, is not possible. This is because the master activation procedure completely skips the process data memory allocation if it has already been previously allocated (or not freed in this case).